// node-projects/backups/nub/20250424/interface/types.ts
export type TStatusQuery = {
  providerId: ProviderId;
  status: AppTxStatus;
  /** Path (in users_transactions_withdrawal collection) to doc containing data related to the utility payment request */
  txPath?: string;
} & { status: AppTxStatus.Mempool; txPath: string };

export enum Utility {
  Mobile = "mobile",
  Electricity = "electricity",
  Cable = "cableTv",
  Bet = "bet",
}
// TODO: update
export enum UtilityMerchant {
  // Mobile merchants
  Mtn = "mtn",
  Airtel = "airtel",
  Glo = "glo",
  NineMobile = "9mobile",
  Spectranet = "spectranet",
  // Electricity merchants
  IkejaElectric = "ikejaElectric",
  AbujaElectric = "abujaElectric",
  EkoElectric = "ekoElectric",
  KadunaElectric = "kadunaElectric",
  AbaElectric = "abaElectric",
  // Cable TV merchants
  Dstv = "dstv",
  Gotv = "gotv",
  Startimes = "startimes",
  // Bet merchants
  Betking = "betking",
  Bet9ja = "bet9ja",
  Betway = "betway",
  Sportybet = "sportybet",
  Nairabet = "nairabet",
}
// TODO: merchantService -> product
export enum UtilityMerchantService {
  Airtime = "airtime",
  Prepaid = "prepaid",
  Postpaid = "postpaid",
  Plan = "plan",
  Default = "default",
}

// TODO: update
export type TPlanData = {
  providerId: ProviderId;
  planCode: string;
  amount: number;
};

export type TPaymentData = TUtilityMeta &
  (
    | {
        // Require plan data as well, if service is of 'plan' type...
        merchantServiceId: UtilityMerchantService.Plan;
        planData: TPlanData;
      }
    | {
        // ...otherwise, it should not be present.
        merchantServiceId: Exclude<UtilityMerchantService, "plan">;
        planData?: undefined;
      }
  );

export interface IUtilityProvider {
  /** Maps merchant product IDs recognized by our app to those of a specific provider. */
  appToProviderProductMap: {
    [K in UtilityMerchant]: { [K2 in UtilityMerchantService]?: string };
  };

  /** Maps merchant IDs recognized by our app to those of a specific provider. */
  appToProviderMerchantMap: { [K in UtilityMerchant]: string };

  /**
   * Determines if a provider supports a merchant service/product. Only used for non-plan services like airtime.
   * For the 'plan' merchant service, provider ID is expected in fetched plan data.
   * @param service - the merchant service ID recognized by our app.
   * @returns true if specified service/product payment is supported
   */
  supports(merchant: UtilityMerchant, service: UtilityMerchantService): boolean;

  /**
   * Make utility payment for airtime.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityMerchantService.Airtime](
    paymentData: TPaymentData
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for a merchant's default service (variable pricing).
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityMerchantService.Default](
    paymentData: TPaymentData
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for merchant services grouped into bundles with fixed prices.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityMerchantService.Plan](
    paymentData: TPaymentData
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for electricity postpaid service.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityMerchantService.Postpaid](
    paymentData: TPaymentData
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for electricity prepaid service.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityMerchantService.Prepaid](
    paymentData: TPaymentData
  ): Promise<IProviderPaymentResult>;

  /** Returns the current transaction status of the bill payment request */
  checkStatus(
    reqQuery: TStatusQuery,
    _reqBody: Record<string, unknown>
  ): Promise<TCheckStatusResult>;

  /**
   * Derive a request reference given a transaction path.
   * @param txPath - full path to transaction doc (users_transactions_withdrawal).
   * @returns reference used to make the payment request to the provider.
   */
  txpathToTxref(txPath: string): string;
  /**
   * Derive a transaction path given a request reference.
   * @param txRef - reference used to make the payment request to the provider.
   * @returns full path to transaction doc (users_transactions_withdrawal).
   */
  txrefToTxpath(txRef: string): string;
}

/** Unique identifier for each of our providers. E.g. vtpass, remita, paga */
export enum ProviderId {
  ProviderX = "providerX", // TODO: remove; test provider
}

export enum ProviderTxStatus {
  Pending = "pending",
  Successfull = "successful",
  Failed = "failed",
}
export enum AppTxStatus {
  Pending = "pending",
  Approved = "approved",
  Rejected = "rejected",
  Mempool = "mempool",
}

export interface IUtilityStatusTxResult {
  statusDocsWritten: boolean;
}

// TODO: add other metadata from the payment request to save in DB
export interface IProviderPaymentResult {
  initialStatus: ProviderTxStatus;
  /** How status update is retrieved; push (webhook) or pull (calling provider) */
  statusUpdateMethods: { isPushToApp: boolean; isPullFromProvider: boolean };
}

export type TStatusDocsWriteTxData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};

/** Result of a provider transaction status check and/or parsing */
export type TCheckStatusResult = { status: AppTxStatus; txPath: string };

// ------------------- MODEL -------------------

/** Contains data about the utility payment request */
export type TUtilityMeta = {
  /** Unique identifier for the customer, at the merchant */
  customerId: string;
  merchantId: UtilityMerchant;
  merchantServiceId: UtilityMerchantService;
  planId?: string;
  amount?: number;
} & (
  | {
      // Require planId if service is of 'plan' type...
      merchantServiceId: UtilityMerchantService.Plan;
      planId: string;
      amount?: undefined;
    }
  | {
      // ...else, require amount instead
      merchantServiceId: Exclude<UtilityMerchantService, "plan">;
      planId?: undefined;
      amount: number;
    }
);